move env configs to support stub file instead of spec file
