+++ spec/spec_helper.rb
@@ -55,6 +55,8 @@ RSpec.configure do |config|
   config.include FactoryBot::Syntax::Methods
   config.include RSpecHtmlMatchers
   config.include StubCurrentUserHelper
+  config.include StubOmniauth
+
   config.mock_with :rspec do |mock_config|
     mock_config.syntax = %i[expect should]
   end
